Split storefront and dashboard APIs
