+++ saleor/dashboard/graphql/mutations.py
@@ -7,8 +7,8 @@ from graphene.types.mutation import MutationOptions
 from graphene_django.form_converter import convert_form_field
 from graphene_django.registry import get_global_registry
 
-from ..utils import get_node
-from .types import Error
+from ...graphql.utils import get_node
+from ...graphql.core.types import Error
 
 registry = get_global_registry()
