+++ saleor/dashboard/urls.py
@@ -1,10 +1,13 @@
+from django.conf import settings
 from django.conf.urls import include, url
+from graphene_django.views import GraphQLView
 
 from . import views as core_views
 from .category.urls import urlpatterns as category_urls
 from .collection.urls import urlpatterns as collection_urls
 from .customer.urls import urlpatterns as customer_urls
 from .discount.urls import urlpatterns as discount_urls
+from .graphql.api import schema
 from .group.urls import urlpatterns as groups_urls
 from .order.urls import urlpatterns as order_urls
 from .product.urls import urlpatterns as product_urls
@@ -21,6 +24,8 @@ urlpatterns = [
     url(r'^products/', include(product_urls)),
     url(r'^customers/', include(customer_urls)),
     url(r'^staff/', include(staff_urls)),
+    url(r'^graphql/', GraphQLView.as_view(
+        schema=schema, graphiql=settings.DEBUG), name='api'),
     url(r'^groups/', include(groups_urls)),
     url(r'^discounts/', include(discount_urls)),
     url(r'^settings/', include(site_urls)),
