+++ tests/test_graphql.py
@@ -2,58 +2,11 @@ import json
 
 import graphene
 import pytest
-from django import forms
+from django.shortcuts import reverse
 
-from saleor.dashboard.category.forms import CategoryForm
-from saleor.graphql.core.mutations import (
-    ModelFormMutation, ModelFormUpdateMutation)
 from saleor.product.models import Category, Product, ProductAttribute
 
-
-class CategoryForm(forms.ModelForm):
-    class Meta:
-        model = Category
-        fields = ['name']
-
-
-def test_model_form_mutation():
-    class CategoryMutation(ModelFormMutation):
-        test_field = graphene.String()
-
-        class Arguments:
-            test_input = graphene.String()
-
-        class Meta:
-            form_class = CategoryForm
-            return_field_name = 'category'
-
-    meta = CategoryMutation._meta
-    assert meta.form_class == CategoryForm
-    assert meta.model == CategoryForm._meta.model
-    assert meta.return_field_name == 'category'
-
-    arguments = meta.arguments
-    # check if declarative arguments are present
-    assert 'test_input' in arguments
-    # check if model form field is present
-    assert 'name' in arguments
-
-    output_fields = meta.fields
-    assert 'category' in output_fields
-    assert 'errors' in output_fields
-
-
-def test_model_form_update_mutation():
-    class CategoryUpdateMutation(ModelFormUpdateMutation):
-        class Meta:
-            form_class = CategoryForm
-
-    meta = CategoryUpdateMutation._meta
-    assert 'id' in meta.arguments
-
-
-def get_content(response):
-    return json.loads(response.content.decode('utf8'))
+from .utils import get_graphql_content
 
 
 def test_category_query(client, product_in_stock):
@@ -88,8 +41,8 @@ def test_category_query(client, product_in_stock):
         }
     }
     ''' % {'category_pk': graphene.Node.to_global_id('Category', category.pk)}
-    response = client.post('/graphql/', {'query': query})
-    content = get_content(response)
+    response = client.post(reverse('api'), {'query': query})
+    content = get_graphql_content(response)
     assert 'errors' not in content
     category_data = content['data']['category']
     assert category_data is not None
@@ -119,8 +72,8 @@ def test_fetch_all_products(client, product_in_stock):
         }
     }
     '''
-    response = client.post('/graphql/', {'query': query})
-    content = get_content(response)
+    response = client.post(reverse('api'), {'query': query})
+    content = get_graphql_content(response)
     assert 'errors' not in content
     num_products = Product.objects.count()
     assert content['data']['products']['totalCount'] == num_products
@@ -142,8 +95,8 @@ def test_fetch_unavailable_products(client, product_in_stock):
         }
     }
     '''
-    response = client.post('/graphql/', {'query': query})
-    content = get_content(response)
+    response = client.post(reverse('api'), {'query': query})
+    content = get_graphql_content(response)
     assert 'errors' not in content
     assert content['data']['products']['totalCount'] == 0
     assert not content['data']['products']['edges']
@@ -183,8 +136,8 @@ def test_product_query(client, product_in_stock):
         }
     }
     ''' % {'category_id': graphene.Node.to_global_id('Category', category.id)}
-    response = client.post('/graphql/', {'query': query})
-    content = get_content(response)
+    response = client.post(reverse('api'), {'query': query})
+    content = get_graphql_content(response)
     assert 'errors' not in content
     assert content['data']['category'] is not None
     product_edges_data = content['data']['category']['products']['edges']
@@ -210,7 +163,7 @@ def test_filter_product_by_category(client, product_in_stock):
     }
     '''
     response = client.post(
-        '/graphql/',
+        reverse('api'),
         {
             'query': query,
             'variables': json.dumps(
@@ -218,7 +171,7 @@ def test_filter_product_by_category(client, product_in_stock):
                     'categoryId': graphene.Node.to_global_id(
                         'Category', category.id)}),
             'operationName': 'getProducts'})
-    content = get_content(response)
+    content = get_graphql_content(response)
     assert 'errors' not in content
     product_data = content['data']['products']['edges'][0]['node']
     assert product_data['name'] == product_in_stock.name
@@ -235,14 +188,14 @@ def test_fetch_product_by_id(client, product_in_stock):
     }
     '''
     response = client.post(
-        '/graphql/',
+        reverse('api'),
         {
             'query': query,
             'variables': json.dumps(
                 {
                     'productId': graphene.Node.to_global_id(
                         'Product', product_in_stock.id)})})
-    content = get_content(response)
+    content = get_graphql_content(response)
     assert 'errors' not in content
     product_data = content['data']['node']
     assert product_data['name'] == product_in_stock.name
@@ -268,8 +221,8 @@ def test_filter_product_by_attributes(client, product_in_stock):
     ''' % {
         'category_id': graphene.Node.to_global_id('Category', category.id),
         'filter_by': filter_by}
-    response = client.post('/graphql/', {'query': query})
-    content = get_content(response)
+    response = client.post(reverse('api'), {'query': query})
+    content = get_graphql_content(response)
     assert 'errors' not in content
     product_data = content['data']['category']['products']['edges'][0]['node']
     assert product_data['name'] == product_in_stock.name
@@ -295,8 +248,8 @@ def test_attributes_query(client, product_in_stock):
         }
     }
     '''
-    response = client.post('/graphql/', {'query': query})
-    content = get_content(response)
+    response = client.post(reverse('api'), {'query': query})
+    content = get_graphql_content(response)
     assert 'errors' not in content
     attributes_data = content['data']['attributes']['edges']
     assert len(attributes_data) == attributes.count()
@@ -322,8 +275,8 @@ def test_attributes_in_category_query(client, product_in_stock):
         }
     }
     ''' % {'category_id': graphene.Node.to_global_id('Category', category.id)}
-    response = client.post('/graphql/', {'query': query})
-    content = get_content(response)
+    response = client.post(reverse('api'), {'query': query})
+    content = get_graphql_content(response)
     assert 'errors' not in content
     attributes_data = content['data']['attributes']['edges']
     assert len(attributes_data) == ProductAttribute.objects.count()
@@ -468,133 +421,3 @@ def test_real_query(client, product_in_stock):
                     'attributesFilter': [filter_by]})})
     content = get_content(response)
     assert 'errors' not in content
-
-
-def test_category_create_mutation(client):
-    query = """
-        mutation($name: String!, $description: String, $parentId: ID) {
-            categoryCreate(
-                name: $name
-                description: $description
-                parentId: $parentId
-            ) {
-                category {
-                    id
-                    name
-                    slug
-                    description
-                    parent {
-                        name
-                        id
-                    }
-                }
-                errors {
-                    field
-                    message
-                }
-            }
-        }
-    """
-
-    category_name = 'Test category'
-    category_description = 'Test description'
-
-    # test creating root category
-    variables = json.dumps({
-        'name': category_name, 'description': category_description})
-    response = client.post(
-        '/graphql/', {'query': query, 'variables': variables})
-    content = get_content(response)
-    assert 'errors' not in content
-    data = content['data']['categoryCreate']
-    assert data['errors'] == []
-    assert data['category']['name'] == category_name
-    assert data['category']['description'] == category_description
-    assert not data['category']['parent']
-
-    # test creating subcategory
-    parent_id = data['category']['id']
-    variables = json.dumps({
-        'name': category_name, 'description': category_description,
-        'parentId': parent_id})
-    response = client.post(
-        '/graphql/', {'query': query, 'variables': variables})
-    content = get_content(response)
-    assert 'errors' not in content
-    data = content['data']['categoryCreate']
-    assert data['errors'] == []
-    assert data['category']['parent']['id'] == parent_id
-
-
-def test_category_update_mutation(client, default_category):
-    query = """
-        mutation($id: ID, $name: String!, $description: String) {
-            categoryUpdate(
-                id: $id
-                name: $name
-                description: $description
-            ) {
-                category {
-                    id
-                    name
-                    description
-                    parent {
-                        id
-                    }
-                }
-                errors {
-                    field
-                    message
-                }
-            }
-        }
-    """
-    # create child category and test that the update mutation won't change
-    # it's parent
-    child_category = default_category.children.create(name='child')
-
-    category_name = 'Updated name'
-    category_description = 'Updated description'
-
-    category_id = graphene.Node.to_global_id('Category', child_category.pk)
-    variables = json.dumps({
-        'name': category_name, 'description': category_description,
-        'id': category_id})
-    response = client.post(
-        '/graphql/', {'query': query, 'variables': variables})
-    content = get_content(response)
-    assert 'errors' not in content
-    data = content['data']['categoryUpdate']
-    assert data['errors'] == []
-    assert data['category']['id'] == category_id
-    assert data['category']['name'] == category_name
-    assert data['category']['description'] == category_description
-
-    parent_id = graphene.Node.to_global_id('Category', default_category.pk)
-    assert data['category']['parent']['id'] == parent_id
-
-
-def test_category_delete_mutation(client, default_category):
-    query = """
-        mutation($id: ID!) {
-            categoryDelete(id: $id) {
-                category {
-                    name
-                }
-                errors {
-                    field
-                    message
-                }
-            }
-        }
-    """
-    variables = json.dumps({
-        'id': graphene.Node.to_global_id('Category', default_category.id)})
-    response = client.post(
-        '/graphql/', {'query': query, 'variables': variables})
-    content = get_content(response)
-    assert 'errors' not in content
-    data = content['data']['categoryDelete']
-    assert data['category']['name'] == default_category.name
-    with pytest.raises(Category.DoesNotExist):
-        default_category.refresh_from_db()
