+++ tests/utils.py
@@ -1,3 +1,4 @@
+import json
 from urllib.parse import urlparse
 
 from django.db.models import Q
@@ -20,3 +21,7 @@ def filter_products_by_attribute(queryset, attribute_id, value):
     in_product = Q(attributes__contains={key: value})
     in_variant = Q(variants__attributes__contains={key: value})
     return queryset.filter(in_product | in_variant)
+
+
+def get_graphql_content(response):
+    return json.loads(response.content.decode('utf8'))
